feat(reader-settings): tabbed reader sheet from reusable Formatting/Display/Behavior sections
Split the monolithic FormattingPanel into three stateless section composables
(Formatting/Display/Behavior) plus shared controls, consumed by a fixed-height
tabbed in-reader sheet. Theme + reading mode + double-page move to Display; the
reader shows a read-only Auto-schedule summary (editing stays in Settings).
