+++ app/src/main/kotlin/com/riffle/app/feature/settings/ReaderSettingsDetailPanels.kt
@@ -1,0 +1,85 @@
+package com.riffle.app.feature.settings
+
+import androidx.activity.compose.BackHandler
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.navigationBarsPadding
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.statusBarsPadding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.riffle.app.feature.reader.BehaviorSection
+import com.riffle.app.feature.reader.DisplaySection
+import com.riffle.app.feature.reader.FormattingSection
+import com.riffle.core.domain.FormattingPreferences
+
+/**
+ * Global Settings drill-in screens for the three reader-settings categories. Each is a full-screen
+ * panel (mirroring [ListeningPreferencesPanel]) wrapping the shared section composable. The Display
+ * panel passes `scheduleEditable = true` so the Auto day/night schedule editor is shown here.
+ */
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+private fun DetailScaffold(title: String, onDismiss: () -> Unit, content: @Composable () -> Unit) {
+    BackHandler(onBack = onDismiss)
+    Surface(modifier = Modifier.fillMaxSize().statusBarsPadding(), tonalElevation = 1.dp) {
+        Column {
+            TopAppBar(
+                title = { Text(title) },
+                navigationIcon = {
+                    IconButton(onClick = onDismiss) {
+                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
+                    }
+                },
+            )
+            Column(
+                modifier = Modifier
+                    .verticalScroll(rememberScrollState())
+                    .padding(horizontal = 24.dp, vertical = 16.dp)
+                    .navigationBarsPadding(),
+            ) { content() }
+        }
+    }
+}
+
+@Composable
+fun FormattingSettingsPanel(
+    prefs: FormattingPreferences,
+    onPrefsChange: (FormattingPreferences) -> Unit,
+    onDismiss: () -> Unit,
+) = DetailScaffold("Formatting", onDismiss) { FormattingSection(prefs, onPrefsChange) }
+
+@Composable
+fun DisplaySettingsPanel(
+    prefs: FormattingPreferences,
+    onPrefsChange: (FormattingPreferences) -> Unit,
+    onDismiss: () -> Unit,
+) = DetailScaffold("Display", onDismiss) { DisplaySection(prefs, onPrefsChange, scheduleEditable = true) }
+
+@Composable
+fun BehaviorSettingsPanel(
+    keepScreenOn: Boolean,
+    onKeepScreenOnChange: (Boolean) -> Unit,
+    volumeKeyNavigationEnabled: Boolean,
+    onVolumeKeyNavigationEnabledChange: (Boolean) -> Unit,
+    invertVolumeKeys: Boolean,
+    onInvertVolumeKeysChange: (Boolean) -> Unit,
+    onDismiss: () -> Unit,
+) = DetailScaffold("Behavior", onDismiss) {
+    BehaviorSection(
+        keepScreenOn, onKeepScreenOnChange,
+        volumeKeyNavigationEnabled, onVolumeKeyNavigationEnabledChange,
+        invertVolumeKeys, onInvertVolumeKeysChange,
+    )
+}
